CompPASS

CompPASS Comparative Proteomic Analysis Software Suite (CompPASS) is based on spoke model. This algorithm was developed by Dr. Mathew Sowa for defining the human deubiquitinating enzyme interaction landscape (Sowa, Mathew E., et al., 2009). The implementation of this algorithm was inspired by Dr. Sowa's online tutorial (http://besra.hms.harvard.edu/ipmsmsdbs/cgi-bin/tutorial.cgi). The output includes Z-score, S-score, D-score and WD-score. This function also computes entropy and normalized WD-score. The source code for this function was based on the source code. https://github.com/dnusinow/cRomppass

CompPASS(datInput)

Arguments

datInput

A dataframe with column names: idRun, idBait, idPrey, countPrey. Each row represent one unique protein captured in one pull-down experiment

Value

A data frame consists of unique bait-prey pairs with Z-score, S-score,D-score and WD-score indicating interacting probabilities.

References

Sowa, Mathew E., et al. "Defining the human deubiquitinating enzyme interaction landscape." Cell 138.2 (2009): 389-403. https://doi.org/10.1016/j.cell.2009.04.042

Huttlin, Edward L., et al. "The BioPlex network: a systematic exploration of the human interactome." Cell 162.2 (2015): 425-440. https://doi.org/10.1016/j.cell.2015.06.043

Huttlin, Edward L., et al. "Architecture of the human interactome defines protein communities and disease networks." Nature 545.7655 (2017): 505. https://www.nature.com/articles/nature22366

Author

Qingzhou Zhang, zqzneptune@hotmail.com

Examples

data(TestDatInput)
datScore <- CompPASS(TestDatInput)
head(datScore)
#>   idBait idPrey AvePSM    scoreZ    scoreS    scoreD Entropy   scoreWD
#> 1  AIFM3  AIFM1     20 7.9382230 36.055513 36.055513       0 1.6903085
#> 2  AIFM3  ALDOA     14 2.6586313  9.095453  9.095453       0 0.2308028
#> 3  AIFM3 ATP5A1      5 0.5826082  5.700877  5.700877       0 0.1529845
#> 4  AIFM3   CALR      4 0.8703043  4.654747  4.654747       0 0.1161689
#> 5  AIFM3   CCT2     24 3.1558989 12.489996 12.489996       0 0.3398374
#> 6  AIFM3   CCT4     20 2.8371693  9.013878  9.013878       0 0.2135599