Hart

Hart Scoring algorithm based on a hypergeometric distribution error model (Hart et al.,2007).

Hart(datInput)

Arguments

datInput

A dataframe with column names: idRun, idPrey. Each row represent one unique protein captured in one pull-down experiment.

Value

A dataframe consists of pairwise combindation of preys identified in the input with Hart scores indicating interacting probabilities computed from negative log transformed Hypergeometric test P-values.

References

Hart, G. Traver, Insuk Lee, and Edward M. Marcotte. 'A high-accuracy consensus map of yeast protein complexes reveals modular nature of gene essentiality.' BMC bioinformatics 8.1 (2007): 236. https://doi.org/10.1186/1471-2105-8-236

Author

Qingzhou Zhang, zqzneptune@hotmail.com

Examples

data(TestDatInput)
datScore <- Hart(TestDatInput)
head(datScore)
#>           PPI InteractorA InteractorB Freq TnA TnB totTn     Hart
#> 1  AARS~ABCE1        AARS       ABCE1    1   3   2  5000 15.24283
#> 2 AARS~ACADSB        AARS      ACADSB    1   3   3  5000 14.14395
#> 3  AARS~ACAT1        AARS       ACAT1    1   3   9  5000 11.65744
#> 4  AARS~ACBD3        AARS       ACBD3    1   3   4  5000 13.45053
#> 5  AARS~ACTC1        AARS       ACTC1    1   3  16  5000 10.45160
#> 6  AARS~ACTN2        AARS       ACTN2    1   3   4  5000 13.45053