SAINTexpress-spc

SAINTexpress-spc SAINTexpress scoring using spectral count data (spc mode). Implements the SAINT-MRF model with Generalized Poisson distribution for spectral count-based AP-MS data.

SAINTexpress_spc(
  inter_df,
  prey_df,
  bait_df,
  GO_df = NULL,
  f = 0,
  R = 100L,
  L = 100L
)

Arguments

inter_df

A data frame with columns: ip_id, bait_id, prey_id, quant. Each row represents one interaction observation (spectral count).

prey_df

A data frame with columns: prey_id. Optional columns: prey_length, gene_name.

bait_df

A data frame with columns: ip_id, bait_id, test_ctrl. test_ctrl must be "T" (test) or "C" (control).

GO_df

Optional data frame with columns: go_term, genes. genes should be space-separated gene names.

f

Frequency parameter (0 < f < 1). Default 0.5.

R

Maximum number of replicates for scoring. Default 100.

L

Maximum number of controls per prey. Default 100.

Value

A data frame with columns: Bait, Prey, PreyGene, Spec, SpecSum, AvgSpec, NumReplicates, ctrlCounts, AvgP, MaxP, TopoAvgP, TopoMaxP, SaintScore, logOddsScore, FoldChange, BFDR, boosted_by.

References

Teo G, et al. SAINTexpress: improvements and additional features in Significance Analysis of INTeractome software. J Proteomics. 2014;100:37-43.

Author

Hyungwon Choi, Damian Fermin (original SAINTexpress), Qingzhou Zhang (Rcpp integration)

Examples

bait_path <- system.file("exdata", "TIP49", "bait.dat", package = "SMAD")
prey_path <- system.file("exdata", "TIP49", "prey.dat", package = "SMAD")
inter_path <- system.file("exdata", "TIP49", "inter.dat", package = "SMAD")

bait <- read.table(bait_path, sep = "\t", header = FALSE, 
                   col.names = c("ip_id", "bait_id", "test_ctrl"))
prey <- read.table(prey_path, sep = "\t", header = FALSE, 
                   col.names = c("prey_id", "prey_length"))
inter <- read.table(inter_path, sep = "\t", header = FALSE, 
                    col.names = c("ip_id", "bait_id", "prey_id", "quant"))

result <- SAINTexpress_spc(inter, prey, bait)
head(result)
#>    Bait   Prey PreyGene Spec SpecSum AvgSpec NumReplicates          ctrlCounts
#> 1 ACTR5  ACTR5    ACTR5    0       0       0             1   0|0|0|0|0|0|0|0|0
#> 2 ACTR5 RUVBL2   RUVBL2   73      73      73             1   6|0|1|0|0|0|0|0|0
#> 3 ACTR5 RUVBL1   RUVBL1   50      50      50             1   3|2|0|0|0|0|0|2|0
#> 4 ACTR5 INO80C   INO80C   26      26      26             1   0|0|0|0|0|0|0|0|0
#> 5 ACTR5  ACTR8    ACTR8   32      32      32             1   0|0|0|0|0|0|0|0|0
#> 6 ACTR5   CCT2     CCT2   16      16      16             1 1|14|20|0|0|0|0|0|0
#>        AvgP      MaxP  TopoAvgP  TopoMaxP SaintScore logOddsScore FoldChange
#> 1 0.0000000 0.0000000 0.0000000 0.0000000  0.0000000   -0.2753709   0.000000
#> 2 0.9999999 0.9999999 0.9999999 0.9999999  0.9999999   16.4238298  93.857143
#> 3 1.0000000 1.0000000 1.0000000 1.0000000  1.0000000   32.4403049  64.285714
#> 4 1.0000000 1.0000000 1.0000000 1.0000000  1.0000000   41.2557466 260.000000
#> 5 1.0000000 1.0000000 1.0000000 1.0000000  1.0000000   49.9730196 320.000000
#> 6 0.9328368 0.9328368 0.9328368 0.9328368  0.9328368    2.6311045   4.114286
#>           BFDR boosted_by
#> 1 3.658563e-01           
#> 2 2.994328e-09           
#> 3 2.220446e-16           
#> 4 0.000000e+00           
#> 5 0.000000e+00           
#> 6 8.106743e-03